third_party/blink/web_tests/svg/hittest/foreign-object-with-transform.html

<!DOCTYPE html>
<title>Hit-test on contents of transformed foreignObject</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<svg>
  <foreignObject width="100" height="100" transform="translate(100, 0)">
    <div id="target" style="width: 100px; height: 100px; background-color: blue"></div>
  </foreignObject>
</svg>
<script>
  test(function() {
    assert_equals(document.elementFromPoint(150, 50), target);
  });
</script>
